Better folder listings, support stat result list.
Gave FolderListing it's own type with the stat-like info nested in the
stat field, which has a type of Folder. The listing is in the 'listing'
field, which is defined as a list of FilesystemObjects.

Modified the filesystemStat query to accept multiple paths and return
multiple results. The return value is not a map, it's a list of object
since the map returned by the stat endpoint is
difficult/confusing/impossible to model with GraphQL, AFAIK. This also
sets the stage for us to use Facebook's dataloader to cache results from
the stat endpoint if it becomes necessary.