Adds support for file uploads.
Files can be POSTed to the authenticated /api/upload path using a form.

A test form is available on /upload, which will be updated later for
more functionality.

A snapshot was updated due to updates to dependencies that occurred when
the requests library was added.

To try things out with your local config:
* Pull the PR.
* npm ci
* npm run dev
* Access http://localhost:3000/login in a browser and login
* Access http://localhost:3000/upload in a browser and upload a file
* If successful, check for the file in the DE.

For the temporary test form on /upload, you're currently expected to
login as ipcdev and files land in ipcdev's home directory. That will
not be the case in the near future.