Start of major refactoring.
The getOrigination() call has been split up and added as method on the
dashboard item classes.
JSX inside of constructors is a no-no (obvious in hindsight), so I added
static create methods on the dashboard items that instantiate the item
objects and add the actions once it's created.
Wired up the new dashboard item classes and objects.
Added the actions to the CardActions section of the cards.
Still needs sizing adjustments.