Add new, non-admin VICE API calls for loading page
Adds GET /vice/resources which returns a filtered VICE analysis listing
for the logged in user.
Adds GET /vice/:host/description which returns a VICE analysis listing
entry for the analysis associated with :host. Access permission for the
logged in user is checked.
Adds GET /vice/:host/url-ready which returns whether or not the VICE
analysis associated with :host is ready for access by browsers. Access
permission for the logged in user is checked.