CORE-702 Refactor App Launch Stepper components for App Editor
Refactored the App Launch Stepper and Step content display components
into common components that can also be used by the App Editor.

Also refactored the Step content component from a fieldset to a MUI
Card, which can display navigation buttons and actions in a
CardActions area, or in the CardHeader actions area.