CORE-1369 Add overflow style to PageWrapper
Fixes the App Launch form getting cutoff at the bottom of the page,
and doesn't appear to affect appearance of other components using it.

The 100% width style of the AppInfo title and description
were causing horizontal scrolling with the new overflow style,
but it appears that style is not needed anyway.