CORE-1401 Improve loading display of Reference Genome params
Updated ReferenceGenomeSelect params to use cached query data so that
a loading error is not displayed if another Reference Genome param in
the same form is already fetching.

Also increased the query's stale time so that the list is fetched less
often (since this list is no longer intended to be updated).