CORE-1465 Allow arbitrary max resource limits in app/tool forms
The app and tool resource request/limit select fields will still be
populated with powers of 2, starting from 1, but the configs can set a
max value that is not a power of 2 (such as our actual limits in the
cluster), and that max value will still display as the upper limit in
the selection list.
This also fixes a bug in the Edit Tool form that could display a power
of 2 beyond the actual limit set in the configs.