CORE-1466 Add a final, "Completion" step to the Workflow Editor
Adds a final step with different help messages depending on whether
the form has unsaved changes, errors, or is saved/unchanged.

* The message displayed for a saved/unchanged form includes links for
exiting the form (back) or navigating to this app's launch page.

* The message displayed for unsaved changes also includes 2 new
actions: "Save and Exit" and "Save and Launch".

* Since these new actions are contained, primary color buttons,
in addition to the main `Back` and `Save` buttons, the step navigation
buttons were changed from "contained" to "outlined" variants.

* Also commented out the `push` override in the Storybook
`withNextRouter` decorator, so that `push` logs are displayed in the
Storybook "Actions" tab.

* Also prevented the `dirty` prop in the `BackButton` from being
spread down to its inner Button, which fixes a DOM warning.