Fix unexpected error when refreshing Data page
There might be a race condition between when the user is considered logged in vs out.  The case I saw in the console was where the path was `/cyverse/home/aramsey`, but the `dataRoots` were all set for the anonymous user.  This resulted in the `selectedIndex` being set to `-1` which threw an error in the render function when trying to access `dataRoots[selectedIndex]`.