CORE-1609 Fix persistence error in tag clause
Noticed that if you keep opening and closing the dialog while modifying the list of tags, the state of the dlg gets cleared on close which gets the form state out of sync with formik.  So if you added a tag in one search, then wanted to remove it in a secondary search, it wouldn't work.