CORE-1808 Update app search results to display app types
Updated the app search results to remove the system ID and integrator
columns, replacing them with an app "Type" column.

Refactored the getAppTypeDisplay function out of the apps listing
TableView component and into apps utils.

Also hides the TableSortLabel in the SearchResultsTable component if
the column can not be sorted.