CORE-1939 Add usage percentage to Data nav icon
Updated the AppBar's subscription query to also save the user's data
usage percentage, which is passed to the DrawerItems.

Updated DrawerItems to render the DataIcon wrapped in a badge,
and with an info icon for usage percentages above 50%,
a warning icon for usgage above 75%,
with an error icon for usage above 100%,
and with the percentage shown in a tooltip.

Also added a subscriptions util function for the common usage
percentage format shown in the data and CPU usage charts in the
dashboard, and now in the Data nav icon.