CORE-1936 Attempt to fix hooks exceptions in dashboard items
Sometimes when new analyses are launched and added to the
"Recent Analyses" section of the dashboard, navigating away, then back
to the dashboard displays an error:
`A client-side exception has occurred`

The browser dev console displays
"Minified React error #310;
visit https://reactjs.org/docs/error-decoder.html?invariant=310"

Which translates to
`Rendered more hooks than during the previous render`.

A possible fix for this may have been added to the `InstantLaunchItem`
in #312, so I'm attempting this fix in the remaining dashboard items.