CORE-2002 Add countdown timers with controls to analyses stories
Updated the analyses listing, submission landing, and dashboard
stories to display the running VICE countdown timer,
but only when a time limit is selected in the story controls.

Updated the start and end dates in the `info` analysis mock
to help the "Running for" duration labels to make more sense.