CORE-1985 Disable NotificationsProvider websockets in stories
Added a `wsEnabled` prop (defaults to `true`) to NotificationsProvider
so that websockets can be disabled in stories, which conflict with
Storybook's HMR websockets.

Also fixed the logic for checking if `useNotifications` is called
within a NotificationsProvider by passing no default value to the
`React.createContext` call that creates the `NotificationsContext`.