Refactor Global Alerts out of AppBar and into top-level PageWrapper
Since rendering alert markdown is async, the AppBar's height is used
for setting the top-level PageWrapper's `appBarHeight` prop before
the alert messages are rendered, which caused the bottom of pages to
be cut off.

This refactors rendering the active alerts into a `GlobalAlerts`
component that is nested inside the top-level `PageWrapper`,
just above each page's content.

This will also wrap the alerts in an outer `Paper` component for a
better layout inside the `PageWrapper`.