CORE-2057 Fix Admin EditTool form
When an admin attempted to edit a tool, a null tool error in the
service facade prevented the PATCH request from being submitted.

The `handleSubmit` function was setting the formatted values in a
`toolSubmission` state, then immediately calling `updateCurrentTool`
or `addNewTool` mutation queries, which were trying to access that
state, but the state values were not available yet.